0-hash_table_create.c

+++ hash_tables/0-hash_table_create.c
@@ -2,17 +2,17 @@
 
 /**
  * hash_table_create - create an hash table
- * size: size of the hash table
+ * @size: izesize of the hash table
  * Return: adress of the hash table
  */
 
 hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *htable = NULL;
-	int i;
+	unsigned long int i;
 
 	htable = malloc(sizeof(hash_table_t));
-	if(htable == NULL)
+	if (htable == NULL)
 		return (NULL);
 	htable->array = malloc(sizeof(hash_node_t) * size);
 	if (htable->array == NULL)
@@ -21,7 +21,7 @@ hash_table_t *hash_table_create(unsigned long int size)
 		return (NULL);
 	}
 	htable->size = size;
-	for (i= 0; i < size; i++)
+	for (i = 0; i < size; i++)
 		htable->array[i] = NULL;
 	return (htable);
 }
